feat: Add dynamic import for DownloadBanner in layout and page pages
